requirements.txt
requests
ConfigParser
colorama
argparse
prettytable
gevent
python-nmap